Version 2.1
Do not run :helptags automatically.  The user should install the docs (totd.txt) himself.  The tag for vimtips.txt is still inserted automatically.
Reason:  as long as this is the only script that automatically edits the tags file, there is no problem, but what if someone else tries the same trick?  Each script runs :helptags, wiping out the other's customization, ...